msvc: fix gr_type size_t vector replace
The long type is not guaranteed to be 8 bytes on all 64-bit compilers. In this case, long is 4 bytes and this causes a swig compile error for vector_map.h. This fixes the build by checking the size of long to make the swig %apply the correct size. Tested on msvc 2012 and 64-bit ubuntu.